Move PluginRegistryServiceError to its own file
