+++ src/rust/plugin-registry/src/lib.rs
@@ -4,6 +4,7 @@ use structopt::StructOpt;
 
 pub mod client;
 mod deploy_plugin;
+pub mod error;
 pub mod nomad_cli;
 pub mod nomad_client;
 pub mod server;
